Some changes in file structure
