+++ src/emulator/cpu/instructions.rs
@@ -5,10 +5,10 @@ use crate::emulator::bus::{Bus};
 use std::fmt;
 use std::mem::transmute;
 
-const ZERO_FLAG: u8 = 0;
-const NEGATIVE_FLAG: u8 = 1;
-const HALFCARRY_FLAG: u8 = 2;
-const CARRY_FLAG: u8 = 3;
+pub const ZERO_FLAG: u8 = 0;
+pub const NEGATIVE_FLAG: u8 = 1;
+pub const HALFCARRY_FLAG: u8 = 2;
+pub const CARRY_FLAG: u8 = 3;
 
 #[derive(Clone, Copy)]
 pub struct Instruction {
@@ -49,277 +49,6 @@ impl Instruction {
     
         self.cycles
     }
-
-    //Generic functions appear first 
-
-    //Stack management functions
-    fn push_to_stack(registers: &mut Registers, mem: &mut Bus, short: u16){
-        let sp: u16 = registers.SP(Action::Read).value();
-
-        mem.write_byte(sp.wrapping_sub(1), (short >> 8) as u8); 
-        mem.write_byte(sp.wrapping_sub(2), short as u8);
-
-        registers.SP( Action::Write(sp.wrapping_sub(2)) );
-    }
-
-    fn pop_from_stack(registers: &mut Registers, mem: &mut Bus) -> u16 {
-        let sp: u16 = registers.SP(Action::Read).value();
-        registers.SP(Action::Write(sp+2));
-
-
-        let b1: u8 = mem.read_byte(sp).value();
-        let b2: u8 = mem.read_byte(sp.wrapping_add(1)).value();
-        
-        b1 as u16 | (b2 as u16) << 8 //>
-    }
-
-    //Generic increment function for 8 bit registers. 
-    fn INC(registers: &mut Registers, mut val: u8) -> u8{
-        //will lower nibble overflow?
-        if (val & 0x0F) == 0x0F {
-            registers.set_flag(HALFCARRY_FLAG);
-        } else {
-            registers.clear_flag(HALFCARRY_FLAG);
-        }
-
-        val = val.wrapping_add(1);
-
-        if val == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        return val;
-    }
-
-    //generic decrement function for 8 bit registers. 
-    fn DEC(registers: &mut Registers, mut val: u8) -> u8{
-        //will lower nibble overflow?
-        if (val & 0x0F) == 0x0F {
-            registers.set_flag(HALFCARRY_FLAG);
-        } else {
-            registers.clear_flag(HALFCARRY_FLAG);
-        }
-
-        val = val.wrapping_sub(1);
-
-        if val == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.set_flag(NEGATIVE_FLAG);
-
-        return val;
-    }
-
-    fn ADD_u8(registers: &mut Registers, X: u8, use_carry: bool) {
-
-		let A: u8 = registers.A( Action::Read ).value();
-        let C = (use_carry && registers.test_flag(CARRY_FLAG)) as u8;
-        let result: u8 = A.wrapping_add(X).wrapping_add(C);
-
-		let carry = (A as i16 + X as i16 + C as i16) > 0xFF;
-		let half_carry = (A & 0xF) + (X & 0xF) + C > 0xF;
-        
-        if carry { 
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        if half_carry {
-            registers.set_flag(HALFCARRY_FLAG);
-        } else {
-            registers.clear_flag(HALFCARRY_FLAG);
-        }
-
-        if result == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.clear_flag(NEGATIVE_FLAG);
-
-		registers.A( Action::Write(result as u16)) ;
-    }
-
-    fn SUB_u8(registers: &mut Registers, X: u8, use_carry: bool) {
-
-		let A: u8 = registers.A( Action::Read ).value();
-		let C = (use_carry && registers.test_flag(CARRY_FLAG)) as u8;
-		let result: u8 = A.wrapping_sub(X).wrapping_sub(C);
-		let carry = (A as i16 - X as i16 - C as i16) < 0;
-		let half_carry = (A & 0xF) as i16 - (X & 0xF) as i16 - (C as i16) < 0;
-        
-        if carry { 
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        if half_carry {
-            registers.set_flag(HALFCARRY_FLAG);
-        } else {
-            registers.clear_flag(HALFCARRY_FLAG);
-        }
-
-        if result == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.set_flag(NEGATIVE_FLAG);
-
-		registers.A( Action::Write(result as u16)) ;
-    }
-
-    fn AND_u8(registers: &mut Registers, X: u8) {
-        let A: u8 = registers.A( Action::Read ).value();
-        let result = A & X;
-
-        registers.clear_flag(CARRY_FLAG);
-        registers.set_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        if result == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.A( Action::Write(result as u16) );
-    }
-
-    fn OR_u8(registers: &mut Registers, X: u8) {
-        let A: u8 = registers.A( Action::Read ).value();
-        let result = A | X;
-
-        registers.clear_flag(CARRY_FLAG);
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        if result == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.A( Action::Write(result as u16) );
-    }
-
-    fn XOR_u8(registers: &mut Registers, X: u8) {
-        let A: u8 = registers.A( Action::Read ).value();
-        let result = A ^ X;
-
-        registers.clear_flag(CARRY_FLAG);
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        if result == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.A( Action::Write(result as u16) );
-    }
-
-    fn CP_u8(registers: &mut Registers, X: u8) {
-		let A: u8 = registers.A( Action::Read ).value();
-		Instruction::SUB_u8(registers, X, false);
-        registers.A( Action::Write(A as u16) );
-    }
-
-    fn ADD_u16(registers: &mut Registers, X: u16, Y: u16) -> u16 {    
-        let value: u16 = X.wrapping_add(Y);
-
-        let carry = (0x8000 & X) == 0x8000;
-        let halfcarry = (0x0800 & X) == 0x0800;
-
-        let new_carry = (0x8000 & value) == 0x8000;
-        let new_halfcarry = (0x0800 & value) == 0x0800;
-
-        if carry && !new_carry {
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        if halfcarry && !new_halfcarry {
-            registers.set_flag(HALFCARRY_FLAG);
-        } else {
-            registers.clear_flag(HALFCARRY_FLAG);
-        }
-
-        registers.clear_flag(ZERO_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        return value;
-    }
-
-    fn RR(registers: &mut Registers, mut value: u8, carry: bool, zflag_on: bool) -> u8{
-
-        let bit = value.test_bit(1);
-
-        value = match carry {
-            true => (value >> 1) | registers.test_flag(CARRY_FLAG) as u8, //>
-            false => value.rotate_right(1)
-        };
-
-        if bit {
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        if value == 0 && zflag_on {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        return value;
-    }
-
-    fn RL(registers: &mut Registers, mut value: u8, carry: bool, zflag_on: bool) -> u8{
-
-        let bit = value.test_bit(7);
-
-        value = match carry {
-            true => (value << 1) | registers.test_flag(CARRY_FLAG) as u8, //>
-            false => value.rotate_left(1)
-        };
-
-        if bit {
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-
-        if value == 0 && zflag_on {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        return value;
-    }
-
-
-
 
     //0x00
     pub fn NOP_R(_operands: [u8; 2], _registers: &mut Registers, _mem: &mut Bus ){
@@ -3003,26 +2732,15 @@ impl Instruction {
     }
 
     //0xF6
-    pub fn OR_n(operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        
+    pub fn OR_n(operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {  
         Instruction::OR_u8(registers, operands[0]);
-
-        
-
-        
     }
 
     //0xF7
     pub fn RST_30(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus) {
-
         let PC: u16 = registers.PC( Action::Read ).value(); 
         Instruction::push_to_stack(registers, mem, PC);
         registers.PC( Action::Write(30) );
-
-
-        
-
-        
     }
 
     //0xF8
@@ -3054,967 +2772,33 @@ impl Instruction {
 
     //0xF9
     pub fn LD_SP_HL(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        
         let HL: u16 = registers.HL( Action::Read ).value();
-
         registers.SP( Action::Write(HL) );
-
     }
 
     //0xFA
     pub fn LD_A_dnn(operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        
         let nn: u16 = Bus::to_short(operands);
-
         let dnn: u8 = mem.read_byte(nn).value();
-
         registers.A( Action::Write(dnn as u16) );
-
     }
 
     //0xFB
-    pub fn EI(_operands: [u8; 2], _registers: &mut Registers, mem: &mut Bus) {
-        
+    pub fn EI(_operands: [u8; 2], _registers: &mut Registers, mem: &mut Bus) { 
         mem.enable_interrupts();
-
-
-        
-
-        
     }
 
     //0xFE
     pub fn CP_n(operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-
         Instruction::CP_u8(registers, operands[0]);
-
-        
-
-        
     }
 
     //0xF7
     pub fn RST_38(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus) {
-
         let PC: u16 = registers.PC( Action::Read ).value(); 
         Instruction::push_to_stack(registers, mem, PC);
         registers.PC( Action::Write(38) );  
     }
 
-
-    fn SR(byte: u8, preserve: bool, registers: &mut Registers) -> u8 {
-		let shifted = match preserve {
-			true => (byte >> 1) | (byte & 0x80),
-			false => byte >> 1,
-        };
-        
-        if (byte & 1) == 1 {
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        if byte == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-        
-        shifted
-	}
-
-    fn SL(byte: u8, preserve: bool, registers: &mut Registers) -> u8 {
-		let shifted = match preserve {
-			true => (byte << 1) | (byte & 0x01), //>
-			false => byte << 1, //>
-        };
-        
-        if (byte & 0x80) == 0x80 {
-            registers.set_flag(CARRY_FLAG);
-        } else {
-            registers.clear_flag(CARRY_FLAG);
-        }
-
-        if byte == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-        
-        shifted
-	}
-
-
-
-    //generic test bit for zero
-    fn test_bit(value: u8, bit: u8, registers: &mut Registers) {
-        if value.test_bit(bit) {
-            registers.clear_flag(ZERO_FLAG);
-        } else {
-            registers.set_flag(ZERO_FLAG);
-        }
-
-        registers.set_flag(HALFCARRY_FLAG);
-        registers.clear_flag(NEGATIVE_FLAG);
-    }
-
-    fn swap(value: u8, registers: &mut Registers) -> u8 {
-
-        if value == 0 {
-            registers.set_flag(ZERO_FLAG);
-        } else {
-            registers.clear_flag(ZERO_FLAG);
-        }
-
-        let upper = value & 0xF0;
-        let lower = value & 0x0F;
-
-        registers.clear_flag(NEGATIVE_FLAG);
-        registers.clear_flag(HALFCARRY_FLAG);
-        registers.clear_flag(CARRY_FLAG);
-
-        lower << 4 | upper >> 4
-    }
-
-    //
-    //  CB PREFIX INSTRUCTIONS
-    //
-
-    //0xCB 0x00
-    pub fn RLC_B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut B: u8 = registers.B( Action::Read ).value();
-
-        B = Instruction::RL(registers, B, false, true);
-
-        registers.B( Action::Write(B as u16) );
-    }
-
-    //0xCB 0x01
-    pub fn RLC_C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut C: u8 = registers.C( Action::Read ).value();
-
-        C = Instruction::RL(registers, C, false, true);
-
-        registers.C( Action::Write(C as u16) );
-    }
-
-    //0xCB 0x02
-    pub fn RLC_D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut D: u8 = registers.D( Action::Read ).value();
-
-        D = Instruction::RL(registers, D, false, true);
-
-        registers.D( Action::Write(D as u16) );
-    }
-
-    //0xCB 0x03
-    pub fn RLC_E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut E: u8 = registers.E( Action::Read ).value();
-
-        E = Instruction::RL(registers, E, false, true);
-
-        registers.E( Action::Write(E as u16) );
-    }
-
-    //0xCB 0x04
-    pub fn RLC_H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut H: u8 = registers.H( Action::Read ).value();
-
-        H = Instruction::RL(registers, H, false, true);
-
-        registers.H( Action::Write(H as u16) );
-    }
-
-    //0xCB 0x05
-    pub fn RLC_L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut L: u8 = registers.L( Action::Read ).value();
-
-        L = Instruction::RL(registers, L, false, true);
-
-        registers.L( Action::Write(L as u16) );
-    }
-
-    //0xCB 0x06
-    pub fn RLC_dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let mut val = mem.read_byte(dHL).value();
-
-        val = Instruction::RL(registers, val, false, true);
-
-        mem.write_byte(dHL, val);
-    }
-
-    //0xCB 0x07
-    pub fn RLC_A_CB(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A = Instruction::RL(registers, A, false, true);
-
-        registers.A( Action::Write(A as u16) );
-    }
-
-    //0xCB 0x08
-    pub fn RRC_B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.B( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.B( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x09
-    pub fn RRC_C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.C( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.C( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x0A
-    pub fn RRC_D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.D( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.D( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x0B
-    pub fn RRC_E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.E( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.E( Action::Write(val as u16) );
-    }
-    //0xCB 0x0C
-    pub fn RRC_H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.H( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.H( Action::Write(val as u16) );
-    }
-    //0xCB 0x0D
-    pub fn RRC_L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.L( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        registers.L( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x0E
-    pub fn RRC_dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let mut val = mem.read_byte(dHL).value();
-
-        val = Instruction::RR(registers, val, false, true);
-
-        mem.write_byte(dHL, val);
-    }
-
-    //0xCB 0x0F
-    pub fn RRC_A_CB(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A = Instruction::RR(registers, A, false, true);
-
-        registers.A( Action::Write(A as u16) );  
-    }
-
-    //0xCB 0x10
-    pub fn RL_B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.B( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.B( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x11
-    pub fn RL_C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut C: u8 = registers.C( Action::Read ).value();
-
-        C = Instruction::RL(registers, C, true, true);
-
-        registers.C( Action::Write(C as u16) );
-    }
-
-    //0xCB 0x12
-    pub fn RL_D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.D( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.D( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x13
-    pub fn RL_E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.E( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.E( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x14
-    pub fn RL_H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.H( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.H( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x15
-    pub fn RL_L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.L( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.L( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x16
-    pub fn RL_dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let mut val = mem.read_byte(dHL).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        mem.write_byte(dHL, val);
-    }
-
-    //0xCB 0x17
-    pub fn RL_A_CB(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.A( Action::Read ).value();
-
-        val = Instruction::RL(registers, val, true, true);
-
-        registers.A( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x18
-    pub fn RR_B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.B( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.B( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x19
-    pub fn RR_C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.C( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.C( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x1A
-    pub fn RR_D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.D( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.D( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x1B
-    pub fn RR_E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.E( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.E( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x1C
-    pub fn RR_H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.H( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.H( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x1D
-    pub fn RR_L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.L( Action::Read ).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        registers.L( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x1E
-    pub fn RR_dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let mut val = mem.read_byte(dHL).value();
-
-        val = Instruction::RR(registers, val, true, true);
-
-        mem.write_byte(dHL, val);
-    }
-
-    //0xCB 0x1F
-    pub fn RR_A_CB(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A = Instruction::RR(registers, A, true, true);
-
-        registers.A( Action::Write(A as u16) );
-    }
-
-    //0xCB 0x21
-    pub fn SLA_C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.C( Action::Read ).value();
-
-
-        val = Instruction::SL(val, true, registers);
-
-        registers.C( Action::Write(val as u16) );
-    }
-
-
-
-
-
-    //0xCB 0x87
-    pub fn RES_0_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(0);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0x8F
-    pub fn RES_1_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(1);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0x97
-    pub fn RES_2_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(2);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0x9F
-    pub fn RES_3_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(3);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0xA7
-    pub fn RES_4_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(4);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0xAF
-    pub fn RES_5_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(5);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0xB7
-    pub fn RES_6_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(6);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-    //0xCB 0xBF
-    pub fn RES_7_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus) {
-        let mut A: u8 = registers.A( Action::Read ).value();
-
-        A.reset_bit(7);
-
-        registers.A( Action::Write( A as u16 ) );
-    }
-
-    //0xCB 0x37
-    pub fn SWAP_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.A( Action::Read ).value();
-        
-        val = Instruction::swap(val, registers);
-        
-        registers.A( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x38
-    pub fn SRL_B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.B( Action::Read ).value();
-        
-        val = Instruction::SR(val, false, registers);
-        
-        registers.B( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x3F
-    pub fn SRL_A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let mut val: u8 = registers.A( Action::Read ).value();
-        
-        val = Instruction::SR(val, false, registers);
-        
-        registers.A( Action::Write(val as u16) );
-    }
-
-    //0xCB 0x40
-    pub fn BIT_0B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x41
-    pub fn BIT_0C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x42
-    pub fn BIT_0D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x43
-    pub fn BIT_0E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x44
-    pub fn BIT_0H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x45
-    pub fn BIT_0L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x46
-    pub fn BIT_0dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x47
-    pub fn BIT_0A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 0, registers);
-    }
-
-    //0xCB 0x48
-    pub fn BIT_1B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x49
-    pub fn BIT_1C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4A
-    pub fn BIT_1D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4B
-    pub fn BIT_1E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4C
-    pub fn BIT_1H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4D
-    pub fn BIT_1L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4E
-    pub fn BIT_1dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x4F
-    pub fn BIT_1A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 1, registers);
-    }
-
-    //0xCB 0x50
-    pub fn BIT_2B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x51
-    pub fn BIT_2C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x52
-    pub fn BIT_2D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x53
-    pub fn BIT_2E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x54
-    pub fn BIT_2H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x55
-    pub fn BIT_2L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x56
-    pub fn BIT_2dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x57
-    pub fn BIT_2A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 2, registers);
-    }
-
-    //0xCB 0x58
-    pub fn BIT_3B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x59
-    pub fn BIT_3C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5A
-    pub fn BIT_3D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5B
-    pub fn BIT_3E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5C
-    pub fn BIT_3H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5D
-    pub fn BIT_3L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5E
-    pub fn BIT_3dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x5F
-    pub fn BIT_3A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 3, registers);
-    }
-
-    //0xCB 0x60
-    pub fn BIT_4B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x61
-    pub fn BIT_4C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x62
-    pub fn BIT_4D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x63
-    pub fn BIT_4E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x64
-    pub fn BIT_4H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x65
-    pub fn BIT_4L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x66
-    pub fn BIT_4dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x67
-    pub fn BIT_4A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 4, registers);
-    }
-
-    //0xCB 0x68
-    pub fn BIT_5B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x69
-    pub fn BIT_5C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x6A
-    pub fn BIT_5D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x6B
-    pub fn BIT_5E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x6C
-    pub fn BIT_5H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x6D
-    pub fn BIT_5L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    
-    }
-
-    //0xCB 0x6E
-    pub fn BIT_5dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 5, registers);
-    }
-
-    //0xCB 0x6F
-    pub fn BIT_5A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 5, registers);
-    
-    }
-
-    //0xCB 0x70
-    pub fn BIT_6B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x71
-    pub fn BIT_6C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x72
-    pub fn BIT_6D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x73
-    pub fn BIT_6E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x74
-    pub fn BIT_6H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x75
-    pub fn BIT_6L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x76
-    pub fn BIT_6dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x77
-    pub fn BIT_6A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 6, registers);
-    }
-
-    //0xCB 0x78
-    pub fn BIT_7B(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.B( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x79
-    pub fn BIT_7C(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.C( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7A
-    pub fn BIT_7D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.D( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7B
-    pub fn BIT_7E(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.E( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7C
-    pub fn BIT_7H(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.H( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7D
-    pub fn BIT_7L(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.L( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7E
-    pub fn BIT_7dHL(_operands: [u8; 2], registers: &mut Registers, mem: &mut Bus){
-        let dHL: u16 = registers.HL( Action::Read ).value();
-
-        let val = mem.read_byte(dHL).value();
-
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0x7F
-    pub fn BIT_7A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-        let val: u8 = registers.A( Action::Read ).value();
-        Instruction::test_bit(val, 7, registers);
-    }
-
-    //0xCB 0xC7
-    pub fn SET_0A(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-
-        let mut val: u8 = registers.A( Action::Read ).value();
-        
-        val.set_bit(0);
-
-        registers.A( Action::Write(val as u16) );
-    }
-
-    //0xCB 0xF2
-    pub fn SET_6D(_operands: [u8; 2], registers: &mut Registers, _mem: &mut Bus){
-
-        let mut val: u8 = registers.D( Action::Read ).value();
-        
-        val.set_bit(6);
-
-        registers.D( Action::Write(val as u16) );
-    }
-    
 
 }
